refactor: member 관련 패키지 이동
<description>
<issue-number>
