+++ src/main/java/com/necklife/api/web/controller/member/MemberController.java
@@ -13,8 +13,11 @@ import com.necklife.api.web.dto.response.member.*;
 import com.necklife.api.web.support.ApiResponse;
 import com.necklife.api.web.support.ApiResponseGenerator;
 import com.necklife.api.web.support.MessageCode;
-import com.necklife.api.web.usecase.*;
-import com.necklife.api.web.usecase.dto.response.*;
+import com.necklife.api.web.usecase.dto.response.member.DeleteMemberUseCaseResponse;
+import com.necklife.api.web.usecase.dto.response.member.GetMemberDetailUseCaseResponse;
+import com.necklife.api.web.usecase.dto.response.member.GetMemberTokenDetailUseCaseResponse;
+import com.necklife.api.web.usecase.dto.response.member.PostMemberUseCaseResponse;
+import com.necklife.api.web.usecase.member.*;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
