+++ src/main/java/com/necklife/api/web/usecase/dto/response/member/DeleteMemberUseCaseResponse.java
@@ -1,4 +1,4 @@
-package com.necklife.api.web.usecase.dto.response;
+package com.necklife.api.web.usecase.dto.response.member;
 
 import java.time.LocalDateTime;
 import lombok.*;
