+++ src/main/java/com/necklife/api/web/usecase/dto/response/member/PostMemberUseCaseResponse.java
@@ -1,4 +1,4 @@
-package com.necklife.api.web.usecase.dto.response;
+package com.necklife.api.web.usecase.dto.response.member;
 
 import lombok.*;
 
@@ -8,7 +8,7 @@ import lombok.*;
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Builder
-public class GetMemberDetailUseCaseResponse {
+public class PostMemberUseCaseResponse {
 	private Long id;
 	private String email;
 	private String provider;
