+++ src/main/java/com/necklife/api/web/usecase/member/GetMemberDetailUseCase.java
@@ -1,8 +1,8 @@
-package com.necklife.api.web.usecase;
+package com.necklife.api.web.usecase.member;
 
 import com.necklife.api.entity.member.MemberEntity;
 import com.necklife.api.repository.member.MemberRepository;
-import com.necklife.api.web.usecase.dto.response.GetMemberTokenDetailUseCaseResponse;
+import com.necklife.api.web.usecase.dto.response.member.GetMemberDetailUseCaseResponse;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -12,15 +12,20 @@ import org.webjars.NotFoundException;
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
-public class GetMemberTokenDetailUseCase {
+public class GetMemberDetailUseCase {
 
 	private final MemberRepository memberRepository;
 
-	public GetMemberTokenDetailUseCaseResponse execute(Long id) {
+	public GetMemberDetailUseCaseResponse execute(Long id) {
 
 		Optional<MemberEntity> findMember = memberRepository.findById(id);
 		findMember.orElseThrow(() -> new NotFoundException("Member not found"));
+		MemberEntity member = findMember.get();
 
-		return new GetMemberTokenDetailUseCaseResponse(findMember.get().getId());
+		return new GetMemberDetailUseCaseResponse(
+				member.getId(),
+				member.getEmail(),
+				member.getOauthProvider().toString(),
+				member.getStatus().toString());
 	}
 }
