+++ src/main/java/com/necklife/api/web/usecase/member/GetMemberTokenDetailUseCase.java
@@ -1,8 +1,8 @@
-package com.necklife.api.web.usecase;
+package com.necklife.api.web.usecase.member;
 
 import com.necklife.api.entity.member.MemberEntity;
 import com.necklife.api.repository.member.MemberRepository;
-import com.necklife.api.web.usecase.dto.response.GetMemberDetailUseCaseResponse;
+import com.necklife.api.web.usecase.dto.response.member.GetMemberTokenDetailUseCaseResponse;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -12,20 +12,15 @@ import org.webjars.NotFoundException;
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
-public class GetMemberDetailUseCase {
+public class GetMemberTokenDetailUseCase {
 
 	private final MemberRepository memberRepository;
 
-	public GetMemberDetailUseCaseResponse execute(Long id) {
+	public GetMemberTokenDetailUseCaseResponse execute(Long id) {
 
 		Optional<MemberEntity> findMember = memberRepository.findById(id);
 		findMember.orElseThrow(() -> new NotFoundException("Member not found"));
-		MemberEntity member = findMember.get();
 
-		return new GetMemberDetailUseCaseResponse(
-				member.getId(),
-				member.getEmail(),
-				member.getOauthProvider().toString(),
-				member.getStatus().toString());
+		return new GetMemberTokenDetailUseCaseResponse(findMember.get().getId());
 	}
 }
