+++ src/main/java/com/necklife/api/web/usecase/member/PostBasicMemberUseCase.java
@@ -1,8 +1,8 @@
-package com.necklife.api.web.usecase;
+package com.necklife.api.web.usecase.member;
 
 import com.necklife.api.repository.member.dto.response.PostMemberRepoResponse;
 import com.necklife.api.service.BasicUserService;
-import com.necklife.api.web.usecase.dto.response.PostMemberUseCaseResponse;
+import com.necklife.api.web.usecase.dto.response.member.PostMemberUseCaseResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
